requirements.txt
# ── Core ────────────────────────────────────────────────────────────────────
streamlit>=1.35.0
pandas>=2.2.0
numpy>=1.26.0
requests>=2.31.0
python-dotenv>=1.0.1

# ── Data ────────────────────────────────────────────────────────────────────
yfinance>=0.2.54
openbb>=4.3.0

# ── AI / LLM ────────────────────────────────────────────────────────────────
groq>=1.0.0               # free LLM (Llama 3.3 70B) — primary
anthropic>=0.27.0         # optional fallback if you add Anthropic credits

# ── Utils ────────────────────────────────────────────────────────────────────
httpx>=0.27.0
